move instantiate_env method
